feat: implemented final changes
