+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+### Committing
+
+To have nice commit messages throughout, we use [commitizen](https://github.com/commitizen/cz-cli#making-your-repo-commitizen-friendly). To commit, use `git cz`.
+
+We require reviewed pull requests before merging to `main`. This means:
+
+1. You cannot directly push to the `main` branch.
+2. You should create a new branch for a change you want to make.
+3. Changes should be as small as possible to allow for a quick review.
+4. Once you have applied all the changes you wanted to make to your branch, you need to open a pull request on GitHub.
+5. You need to assign a reviewer and get approval before you are able to merge your changes.
+6. Your pull request needs to pass all checks on GitHub before being merged.
