+++ hierarchical-image-explorer/public/index.html
@@ -6,7 +6,7 @@
 
 	<title>HIE</title>
 
-	<link rel='icon' type='image/png' href='/favicon.png'>
+	<link rel='icon' type='image/png' href='./favicon.png'>
 	<link rel='stylesheet' href='./global.css'>
 	<link rel='stylesheet' href='./build/bundle.css'>
 	<link href='https://fonts.googleapis.com/css?family=Quicksand' rel='stylesheet'>
