+++ hierarchical-image-explorer/src/config.ts
@@ -1,6 +1,6 @@
 import type { SettingsObject } from './types';
 
-export const DEFAULT_NUM_COLUMNS = 10;
+export const DEFAULT_NUM_COLUMNS = 6;
 export const SERVER_ADDRESS = 'http://nemesis.informatik.uni-ulm.de/main/';
 export const DEFAULT_SETTINGS: SettingsObject = {
   columns: DEFAULT_NUM_COLUMNS,
